requirements.txt
autopep8==1.4.4
certifi==2020.11.8
chardet==3.0.4
Django==2.2.13
django-adminlte2==0.4.3
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-oauth-toolkit==1.3.3
django-tables2==2.3.3
django-tables2-reports==0.1.3
idna==2.10
oauthlib==3.1.0
pep8==1.7.1
pycodestyle==2.5.0
pytz==2018.5
requests==2.25.0
sqlparse==0.2.4
urllib3==1.26.2
